+++ src/part3-webserver/breakpoints/5-ref-and-hashmap/ws.ts
@@ -1,4 +1,4 @@
-import { Effect, Layer } from "effect";
+import { Effect, HashMap, Layer, Option, Ref } from "effect";
 import { CurrentConnections, WSSServer } from "./shared";
 import * as M from "./model";
 import * as S from "@effect/schema/Schema";
@@ -7,9 +7,10 @@ export const Live = Layer.effectDiscard(
   Effect.gen(function* (_) {
     const wss = yield* _(WSSServer);
     const currentConnections = yield* _(CurrentConnections);
+    const connections = yield* _(Ref.get(currentConnections));
     function broadcastMessage(message: M.ServerOutgoingMessage) {
       const messageString = JSON.stringify(message);
-      currentConnections.forEach((conn) => {
+      HashMap.forEach(connections, (conn) => {
         conn._rawWS.send(messageString);
       });
     }
@@ -27,30 +28,33 @@ export const Live = Layer.effectDiscard(
           switch (parsedMessage._tag) {
             case "startup": {
               const { color, name } = parsedMessage;
-              if (!M.colors.includes(color) || currentConnections.has(name)) {
+              if (!M.colors.includes(color) || HashMap.has(connections, name)) {
                 ws.close(); // Close the connection if the color is not available or the name is already taken
                 return;
               }
 
               connectionName = name;
               console.log(`New connection: ${name}`);
-              currentConnections.set(name, {
-                _rawWS: ws,
-                name,
-                color,
-                timeConnected: Date.now(),
-              });
+
+              Ref.update(currentConnections, (connections) =>
+                HashMap.set(connections, name, {
+                  _rawWS: ws,
+                  name,
+                  color,
+                  timeConnected: Date.now(),
+                })
+              );
               broadcastMessage({ _tag: "join", name, color });
               break;
             }
             case "message": {
               if (connectionName) {
-                const conn = currentConnections.get(connectionName);
-                if (conn) {
+                const conn = HashMap.get(connections, connectionName);
+                if (Option.isSome(conn)) {
                   broadcastMessage({
                     _tag: "message",
-                    name: conn.name,
-                    color: conn.color,
+                    name: conn.value.name,
+                    color: conn.value.color,
                     message: parsedMessage.message,
                     timestamp: Date.now(),
                   });
@@ -66,14 +70,16 @@ export const Live = Layer.effectDiscard(
 
       ws.on("close", () => {
         if (connectionName) {
-          const conn = currentConnections.get(connectionName);
-          if (conn) {
+          const conn = HashMap.get(connections, connectionName);
+          if (Option.isSome(conn)) {
             broadcastMessage({
               _tag: "leave",
-              name: conn.name,
-              color: conn.color,
+              name: conn.value.name,
+              color: conn.value.color,
             });
-            currentConnections.delete(connectionName);
+            Ref.update(currentConnections, (connections) =>
+              HashMap.remove(connections, connectionName)
+            );
             console.log(`Connection closed: ${connectionName}`);
           }
         }